+++ game/Player.h
@@ -23,7 +23,7 @@ public:
 
     void move(Vector2&& new_position);
     void fire_weapon_equiped(Map& map, const Vector2& mouse_pointer);
-    void get_damage(float damage);
+    void get_damage(uint8_t damage);
     virtual PlayerImage get_player_image() = 0;
     player_id_t get_id() { return id; }
 
@@ -33,8 +33,8 @@ private:
 protected:  // Por ahora lo dejamo asi
     player_id_t id;
     Equipement equipement;
-    int health;
-    int points;
+    uint8_t health;
+    uint8_t points;
 };
 
 #endif  // GAME_PLAYER_H_
